+++ src/soundboard.js
@@ -1,5 +1,5 @@
-import * as MODULE from "../MaterialKeys.js";
-import {launchpad} from "../MaterialKeys.js";
+import {moduleName,launchpad} from "../MaterialKeys.js";
+import {compatibleCore} from "./misc.js";
 
 export class SoundboardControl{
     constructor(){
@@ -14,19 +14,19 @@ export class SoundboardControl{
         const row = 8-Math.floor(key/10);
         const soundNr = column+row*8;
         
-        const mode = game.settings.get(MODULE.moduleName,'soundboardSettings').mode[soundNr];
+        const mode = game.settings.get(moduleName,'soundboardSettings').mode[soundNr];
         
         if (state == false){
             //if 'hold'
             if (mode == 2)
-                this.playSound(soundNr,false,false);
+                this.prePlaySound(soundNr,false,false);
         }
         else {
             let repeat = false;
             if (mode > 0) repeat = true;
             let play = false;
             if (this.activeSounds[soundNr] == false) play = true;
-            this.playSound(soundNr,repeat,play);
+            this.prePlaySound(soundNr,repeat,play);
         }
     }
 
@@ -37,11 +37,11 @@ export class SoundboardControl{
         for (let i=0; i<64; i++){
             let mode = 0;
             
-            let color = game.settings.get(MODULE.moduleName,'soundboardSettings').colorOff[i];
+            let color = game.settings.get(moduleName,'soundboardSettings').colorOff[i];
             
             if (this.activeSounds[i] != false){
-                mode = game.settings.get(MODULE.moduleName,'soundboardSettings').toggle[i];
-                if (mode == 0) color = game.settings.get(MODULE.moduleName,'soundboardSettings').colorOn[i];
+                mode = game.settings.get(moduleName,'soundboardSettings').toggle[i];
+                if (mode == 0) color = game.settings.get(moduleName,'soundboardSettings').colorOn[i];
             }  
                 
             const row = 8-Math.floor(i/8);
@@ -52,8 +52,8 @@ export class SoundboardControl{
         launchpad.updateLEDs(); 
     }
 
-    async playSound(soundNr,repeat,play){  
-        const soundBoardSettings = game.settings.get(MODULE.moduleName,'soundboardSettings');
+    async prePlaySound(soundNr,repeat,play){  
+        const soundBoardSettings = game.settings.get(moduleName,'soundboardSettings');
         let playlistId;
         if (soundBoardSettings.selectedPlaylists != undefined) playlistId = soundBoardSettings.selectedPlaylists[soundNr];
         let src;
@@ -72,14 +72,14 @@ export class SoundboardControl{
         }
         else {
             const soundId = soundBoardSettings.sounds[soundNr];
-            const sounds = game.playlists.entities.find(p => p._id == playlistId).data.sounds;
+            const sounds = game.playlists.get(playlistId).sounds;
             if (sounds == undefined) return;
-            const sound = sounds.find(p => p._id == soundId);
+            const sound = compatibleCore("0.8.1") ? sounds.find(p => p.id == soundId) : sounds.find(p => p._id == soundId);
             if (sound == undefined) return;
             src = sound.path;
         }
    
-        let volume = game.settings.get(MODULE.moduleName,'soundboardSettings').volume[soundNr]/100;
+        let volume = game.settings.get(moduleName,'soundboardSettings').volume[soundNr]/100;
         volume = AudioHelper.inputToVolume(volume);
         
         let payload = {
@@ -92,21 +92,39 @@ export class SoundboardControl{
         };
         game.socket.emit(`module.MaterialKeys`, payload);
 
-        if (play){
-            volume *= game.settings.get("core", "globalInterfaceVolume");
+        this.playSound(soundNr,src,play,repeat,volume);
+    }
 
-            let howl = new Howl({src, volume, loop: repeat, onend: (id)=>{
-                if (repeat == false){
+    async playSound(soundNr,src,play,repeat,volume){
+        if (play){
+            volume *= game.settings.get("core", "globalAmbientVolume");
+
+            if (compatibleCore("0.8.1")) {
+                let newSound = new SoundNode(src);
+                if(newSound.loaded == false) await newSound.load({autoplay:true});
+                newSound.on('end', ()=>{
+                    if (repeat == false) {
+                        this.activeSounds[soundNr] = false;
+                        this.update();
+                    }
+                });
+                newSound.play({loop:repeat,volume:volume});
+                this.activeSounds[soundNr] = newSound;
+            }
+            else {
+                let howl = new Howl({src, volume, loop: repeat, onend: (id)=>{
+                    if (repeat == false){
+                        this.activeSounds[soundNr] = false;
+                        this.update();
+                    }
+                },
+                onstop: ()=>{
                     this.activeSounds[soundNr] = false;
                     this.update();
-                }
-            },
-            onstop: (id)=>{
-                this.activeSounds[soundNr] = false;
-                this.update();
-            }});
-            howl.play();
-            this.activeSounds[soundNr] = howl;
+                }});
+                howl.play();
+                this.activeSounds[soundNr] = howl;
+            }
         }
         else {
             this.activeSounds[soundNr].stop();